Multi-Build generators correct bundle
In the case where a generator has multiple release types
in one cmake directory, the win and lin bundle incorrectly
had a DEBUG or RELEASE inserted. This corrects that problem
and houses the VST3 in a debug/ or release/ directory

On macOS the built in cmake bundle extensions will handle this.