Linux Timer and FD support (when VST3 Runloop Available) (#85)
This implements linux timer and FD support, and fires the
linux onIdle loop in the VST3 host. It accomplishes this
by tapping into the Steinberg::Linux::IRunLoop which only
exists on the pluginframe. This means absent a UI you currently
don't get linux timer and fd events.  But since the primary use
of these features to date is UI updates, it makes a large number
of claps work in the vst3 wrapper in reaper and bitwig on initial
tests.

Separately from this commit, we should decide what we want to
do about FD polling and timers absent a UI, probably by
consulting with the bitwig and reaper native clap implementations.