A rudimentary standalone wrapper
This commit provides a rudimentary standalone wrapper. The wrapper
makes the following choices

1. It uses rtaudio/rtmidi for auido and midi support
2. It is a Cocoa (mac) GTK3 (lin) and just-exe-until-timo-gets-it
   (win) applications
3. The application binds all midi and default audio at 48k with
   no user choices
4. On mac and lin the application shows the gui if present
5. The commit adds cmake rules to decorate an exeecutable as
   as standalone with clap_entry either statically linked

There are myriad things to do to make this a production worth
standalone, including state, saving options, choosing audio,
a win gui, a timer loop on linux which isn't laughable, an fd loop
on linux, and much more. But this at least puts the stake in the
ground about project cmake and directory structure and
actuall does work with conduit and clap saw demo.