Breakup the cmake
This PR breaks up the cmake into more manageable chunks and factors lots of things into guarantee_
methods to make sure preconditions are met only when needed. it also removes a lot of the CACHE SET
stuff we had to do for subordinate cmake.

It also adds some rudimentary documentation to help navigate the new files.

It also adds an XCode stage to the CI build; and builds the Standalone in CI on all platforms
in CPM download mode