Cleanup the GTK Timers and FD some
The GTK timer implementation was a mess and was inlined in
standalone_host.cpp. No reason we can't deletgate it all to
GtkUtil though, so this commit does that. It also imlements the
timers more natively.  And it uses the glib fd wrapper to
do the posix fd callback, but for some reason this results in
the clap saw demo / vstgui window not even opening. So some work
to do there. Perhaps I need an intermediate widget?

But progress and cleaner code so suggesting we merge this even
though we are still timer-only for working UIs on linux