Supress rtaudio ctest targets
rtaudio cmake incorrectly includes ctest when sub-module
testing is off. https://github.com/thestk/rtaudio/pull/413/files
will fix their module but until it does, this set of BUILD_TESTING
to off will supress the CTest targets

If we end up using CTest in this project before the rtaudio
fix is in place, we will have to think about how to handle this.

Closes #145