Correct a numBuffers/numChannels error
AU sends is for sterep 2 buffers each with one channel, which is
fine, and we initialized our info correctly, but in the process
loop we reset the number of output channels to the channel for the
first buffer. This resulted in always sending mono channel data into
the plugins, which caused conduit to not work.

With this change, auval works and conduit polysynth plays sound in
logic.