Correctly put the bundle version -> auv2 version
We have a couple of versions. BUNDLE_VERSION, PROJECR_VERSION
and the version in the CLAP. BUNDLE_VERSION and PROJECT_VERSION
are required to be digits as is the auv2 version. The clap version
can be anything and may not be projectable to a 32 bit int.

So this makes two changes

1. The BUNDLE_VERSION defaults to the PROJECT_VERSION if not set.
2. The BUNDLE_VERSION becomes the version of the auv2 in the info.plist

We may want to further revisit this, but the hard coded '1' is the
worst choice, so it improves that.