fsutil: use fs::path for loading libraries to fix win32 unicode paths
Previously we were using the ASCII versions of some win32 functions.
These will cause problems if a filesystem path contains non-English
characters.

It made sense to change the Library::load() argument from a const char *
to a fs::path for 2 reasons: it ensures paths are kept in the native
format (utf16-LE for windows, utf8 for linux/mac), and it avoids the
need to do conversions from fs::path to const char * since we were
already using fs::path.