Deal with user specified INSTRUMENT_TYPE
We switch on INSTRUMENT_TYPE to set the internall wrapper
handler. We actually only use this in one spot internally
but we were failing to make a buildable wrapper if you
set the type to something not in our list-of-three.
So for now, warn and default out to internal instrument
for the other types (which is probably the right behavior
for aumf, which is what raised this issue in #271)