Support alternate AUv2 Stream Dictionary Construction Strategies
The AUv2 protocol uses a CFDictionary to constrct its stream.
The default wrapper approach uses code from the AUSDK example and
is a common approach, but not the only one. This allows you to
set other strategies. Right now JUCE is the only other strategy
supported.

This means that we can make AUv2 wrappers compatible with streams
from other toolkits and allow mgiration to the wrapper for projects.