Implement Steinberg::IPlugViewContentScaleSupport (#284)
This passes to gui.set_scale. It's pretty critical to make VST3 and CLAP work identicially on bitwig linux.

Also add a defensive guard against malformed claps. In clap in bws resizing a destroyed gui would post a misbehaving error but in the vst3 we would crash absent this guard. Just a defensive check.

Closes #262