VST3 rescans param names on params::request_rescan
request_rescan with CLAP_PARAM_RESCAN_INFO may invalidate
the name, module, or several other flags. The wrapper up
until now just invalidated the value and re-ran the value
to text.

Shortcircuit allows users to rename macros easily and the clap
correctly calls params_rescan which results in bitwig re-displaying
the clap name. The VST3 and AUv2 don't do this.

This commit fixes the VST3 with a partial correction and a comment.
The partical correction is to re-acquire the parameter name
in a RESCAN_INFO case. The comment explains that we do not
re-acquire the module and build the entire tree.

Would appreciate careful review here. It fixes the shortcircuit bug
but this is pretty fundamental code.