Start of a simplified cmake pattern for clap-first plugins (#339)
This allows you to do `make_clapfirst_plugin` compactly
and get what you need. It is still under development and
experimental, but I'd like to do a merge to next with the
commented warning so I can try it on a few more projects
to bang out bugs, oddities, and omissions.