Fix a lifetime problem in parameter cfName when rescan happens (#352)
Rescan resets info on the internal parameter which CFReleases
the associated cfname. This only happens on rescan.

The parameter info has a cfNameString which is just a copy of
that ref, unretained. The lifetime of that parameter cfstring
needs to be the sweep

Therefore you can

1. Scan and reset the cfstring
2. Which is attached still to a parameter
3. And, occasionally, blammo memory corruption

So instead use the kAudioUnitParameterFlag_CFNameRelease option
and retain the string so *both* the parameter and the AU parameter
info need to be done with it.