Windows: use GET_MODULE_HANDLE_EX_FLAG_UNCHANGED_REFCOUNT to avoid leak
We are using GetModuleHandleEx to get the module handle to the current
module, but without GET_MODULE_HANDLE_EX_FLAG_UNCHANGED_REFCOUNT, the
reference count of the module is incremented. We were never called
FreeLibrary and therefore I believe it leaks. We don't need for the
module handle to stick around since we are just using it to get a few
bits of information immediately.