make getModulePath return fs::path
It returns a file path, so let's use the path type for this.

On Windows, this makes it easy for use to remove the old getModuleName
function which had a few problems:
- Returns a pointer to static memory (not threadsafe, though it
  wouldn't have been a problem ATM I don't think)
- Only allows paths of max length 2048
- Even though it follows the Windows naming pattern, it's better renamed
  to getModulePath since it returns a filepath.

We replace it with a version that requests the exact size and allocates
it into a fs::path which will retain the native wide encoding until we
need to use it/convert it to UTF8.