fix possible WindowsHelper failure
Win32 docs say: "The maximum length for lpszClassName is 256. If
lpszClassName is greater than the maximum length, the RegisterClass
function will fail".

We were using the plugin path which is plausibly longer. Instead, we
already have a unique identifier for the module - the HINSTANCE itself.
So we can just use that.

Also: "window class names need to be unique only within the same
process"
https://learn.microsoft.com/en-us/windows/win32/winmsg/about-window-classes#class-name