Windows filesystem related fixes and minor refactor (#374)
* Windows: use GET_MODULE_HANDLE_EX_FLAG_UNCHANGED_REFCOUNT to avoid leak

We are using GetModuleHandleEx to get the module handle to the current
module, but without GET_MODULE_HANDLE_EX_FLAG_UNCHANGED_REFCOUNT, the
reference count of the module is incremented. We were never called
FreeLibrary and therefore I believe it leaks. We don't need for the
module handle to stick around since we are just using it to get a few
bits of information immediately.

* Move defining of namespace fs to a single location

The same thing was defined in 4 separate places - makes sense to have
a single location.

* make getModulePath return fs::path

It returns a file path, so let's use the path type for this.

On Windows, this makes it easy for use to remove the old getModuleName
function which had a few problems:
- Returns a pointer to static memory (not threadsafe, though it
  wouldn't have been a problem ATM I don't think)
- Only allows paths of max length 2048
- Even though it follows the Windows naming pattern, it's better renamed
  to getModulePath since it returns a filepath.

We replace it with a version that requests the exact size and allocates
it into a fs::path which will retain the native wide encoding until we
need to use it/convert it to UTF8.

* rename getModulePath to getPluginPath

'Plugin' better reflects what it is - the path to the DSO on Windows and
Linux, or the path to the bundle on macOS.

* use fs::path instead of std::filesystem

For consistencies sake essentially.

* remove unnecessary :: with win32 calls

* fix possible WindowsHelper failure

Win32 docs say: "The maximum length for lpszClassName is 256. If
lpszClassName is greater than the maximum length, the RegisterClass
function will fail".

We were using the plugin path which is plausibly longer. Instead, we
already have a unique identifier for the module - the HINSTANCE itself.
So we can just use that.

Also: "window class names need to be unique only within the same
process"
https://learn.microsoft.com/en-us/windows/win32/winmsg/about-window-classes#class-name

* Windows: fix GetModuleFileNameW usage

We were misusing the API, it never returns the actual size, it just
returns what was written into the buffer. So we need to loop and try
again with larger buffers until it fits.

* fix compile error

* fix compile error

* Windows: add max-path limit to GetModuleFileNameW and split to own file

Adds a just-in-case limit to the while-loop that resizes the buffer to
try and fit the path. Additionally:
- Let's start with a buffer size of 150 rather than MAX_PATH. 150 is a
  reasonably large starting point. MAX_PATH is a little deceptive
  because it's not actually relevant in this situation when extended
  paths are involved.
- Since we were doing GetModuleFileNameW in 2 locations and it now has a
  bit of complexity to it, lets split it off into a single function.

* Put osutils_windows in os:: namespace to match generic osutil

* Move getModulePath into header to avoid translation unit issues

Previously it was defined in windows.cpp in the os directory.
Confusingly, this file is VST3-specific; it requires VST3 symbols. To
avoid compiling this file with the windows standalone, let's just make
the little util function header-only. It's only used in 2 places
anyways.

Future TODO: we should separate VST3 specific functionality out of
os/windows.cpp.