Move the clap load out of ApplicationDidFinishLaunching (#378)
That delay for slow loading claps continued to be pesky so
make applicationDidFinishLoading very fast and have it async
call the clap startup code, which can beackball, but at least
activates properly.

The real problem of course is that short circuit debug is too
slow to start! But lots of claps may be.