AU: fix request_callback thread safety, closes #387
In issue #387 we discussed adding a separate queue for communication
between threads other than the audio thread, and the main thread. But
the solution is much simpler.

We just copy what the VST3 wrapper does: set an atomic bool to true and
comsume it in the onIdle callback. Now it is perfectly thread-safe,
simpler and faster.

Previously a message was put into a thread safe queue but it is not
necessary to do that since request_callback has loose timing
requirements meaning we don't need to keep it ordered in a queue.